Refactor plots and dataset creation from Trainer
Trainer did not obey the single responsiblity principle.

* Refactored the plot creation to a PlotManager object in training_plots
* Also refactored the dataset loading code.
